Move route paths out of the routing component
Twenty-eight modules imported the route enums from the component that renders
the router, so linking anywhere pulled in the routing tree. They now live in
their own module, which also lets the bundler group the shared view code by
actual usage instead of around that one hub module.
