+++ client/src/components/KompassiLoginCallback.tsx
@@ -1,6 +1,6 @@
 import { ReactElement, useEffect } from "react";
 import { useNavigate, useSearchParams } from "react-router";
-import { AppRoute } from "client/app/AppRoutes";
+import { AppRoute } from "client/app/routes";
 import { Loading } from "client/components/Loading";
 import { useAppDispatch } from "client/utils/hooks";
 import {
