+++ client/src/markdown/KonstiFaq.mdx
@@ -15,7 +15,7 @@ import { FinnishTime } from "./components/FinnishTime";
 import { Expand } from "./components/Expand";
 import { LotteryThermometerValues } from "./components/LotteryThermometerValues";
 // eslint-disable-next-line @typescript-eslint/no-restricted-imports
-import { ProfileTab } from "../app/AppRoutes";
+import { ProfileTab } from "../app/routes";
 import { ProgramGuide } from "./components/ProgramGuide";
 import { HelpDesks } from "./components/HelpDesks";
 import { PreConventionWeek } from "./components/PreConventionWeek";
