+++ client/src/markdown/KonstiInstructions.mdx
@@ -11,7 +11,7 @@ import { SignupLottery } from "./components/SignupLottery";
 import { Expand } from "./components/Expand";
 import { HelpDesks } from "./components/HelpDesks";
 // eslint-disable-next-line @typescript-eslint/no-restricted-imports
-import { AppRoute, ProgramTab, ProfileTab } from "../app/AppRoutes";
+import { AppRoute, ProgramTab, ProfileTab } from "../app/routes";
 
 <Fi>
 ## Käyttöohjeet
