+++ client/src/views/event-log/EventLogProgramItemMoved.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from "react-i18next";
 import { Link } from "react-router";
 import { EventLogItem } from "shared/types/models/eventLog";
 import { getShortWeekdayAndTime } from "shared/utils/timeFormatter";
-import { AppRoute } from "client/app/AppRoutes";
+import { AppRoute } from "client/app/routes";
 import { useAppSelector } from "client/utils/hooks";
 
 interface Props {
