+++ client/src/views/my-program-items/components/DirectSignupItem.tsx
@@ -5,7 +5,7 @@ import { useNavigate } from "react-router";
 import styled from "styled-components";
 import { isStartTimeChanged } from "shared/utils/isStartTimeChanged";
 import { getShortWeekdayAndTime } from "shared/utils/timeFormatter";
-import { AppRoute } from "client/app/AppRoutes";
+import { AppRoute } from "client/app/routes";
 import { InfoText, InfoTextVariant } from "client/components/InfoText";
 import { SignupQuestionAnswer } from "client/components/SignUpQuestionAnswer";
 import { TertiaryButton } from "client/components/TertiaryButton";
