+++ client/src/views/program-item/signup/components/AdmissionTicketLink.tsx
@@ -2,7 +2,7 @@ import { ReactElement } from "react";
 import { useTranslation } from "react-i18next";
 import { NavLink } from "react-router";
 import styled from "styled-components";
-import { AppRoute } from "client/app/AppRoutes";
+import { AppRoute } from "client/app/routes";
 import { ButtonStyle } from "client/components/Button";
 import { ProgramItemButton } from "client/views/program-item/components/ProgramItemButton";
